requirements.txt
appdirs==1.4.3
distlib==0.3.0
Django==3.0.5
django-appconf==1.0.4
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-storages==1.9.1
filelock==3.0.12
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
importlib-metadata==1.5.0
importlib-resources==1.4.0
keyring==10.6.0
keyrings.alt==3.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
netifaces==0.10.4
oauth==1.0.1
pilkit==2.0
Pillow==5.1.0
protobuf==3.0.0
psycopg2==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.1
urllib3==1.22
virtualenv==20.0.13
